+++ Sources/ApiKit/Integrations/TheMovieDb/TheMovieDb.swift
@@ -1,0 +1,18 @@
+//
+//  TheMovieDb.swift
+//  ApiKit
+//
+//  Created by Daniel Saidi on 2023-03-28.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This namespace contains a TMDB-specific set of environments,
+ routes and data types.
+ 
+ You can setup an api account at `https://themoviedb.org` to
+ start using these types with ApiKit.
+ */
+public struct TheMovieDb {}