+++ Sources/ApiKit/Integrations/Yelp/Yelp.swift
@@ -1,0 +1,18 @@
+//
+//  Yelp.swift
+//  ApiKit
+//
+//  Created by Daniel Saidi on 2023-03-28.
+//  Copyright © 2023 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This namespace contains a Yelp-specific set of environments,
+ routes and data types.
+ 
+ You can setup a dev account at `https://yelp.com/developers`
+ to start using these types with ApiKit.
+ */
+public struct Yelp {}